fix: user-events logs exporter - alternate build pattern
