+++ opentelemetry-user-events-logs/CHANGELOG.md
@@ -2,15 +2,16 @@
 
 ## vNext
 
-- Enhanced validation for the provider name in `with_user_event_exporter(provider_name)`:
+### Changed
+
+- Enhanced validation for the provider name in
+  `with_user_event_exporter(provider_name)`:
   - Empty provider names are now disallowed.
 
-- Event header name is changed to use "Log" instead of "LogRecord.EventName". 
+- Event header name is changed to use "Log" instead of "LogRecord.EventName".
 - `ext_dt` and `ext_cloud` structs are flattened.
-  are flattened.
-
-- **BREAKING** Add builder pattern with `ExportOptions` and implement `build_processor` method [256](https://github.com/open-telemetry/opentelemetry-rust-contrib/pull/256)
-  - Removed `with_user_events_exporter` extension method on `LoggerProviderBuilder`.
+- **BREAKING**
+  - Removed `with_user_events_exporter` extension method on   `LoggerProviderBuilder`.
   - Introduced a builder pattern for the user_events exporter which improves configuration flexibility.
 
   Before:
@@ -28,18 +29,14 @@
   
   ```rust
   use opentelemetry_sdk::logs::LoggerProviderBuilder;
-  use opentelemetry_user_events_logs::{build_processor, ExportOptions};
-  let export_options = ExportOptions::builder("myprovider")
-    .build()
-    .unwrap_or_else(|err| {
-      eprintln!("Failed to create export options. Error: {}", err);
-      panic!("exiting due to error during initialization");
-    });
-  let user_event_processor = build_processor(export_options);
-  LoggerProviderBuilder::default()
-    .with_log_processor(user_event_processor)
-    .build();
+  let user_event_processor = opentelemetry_user_events_logs::Processor::builder("myprovider")
+        .build()
+        .expect("Expected to have processor built");
+    LoggerProviderBuilder::default()
+        .with_log_processor(user_event_processor)
+        .build()
   ```
+
 ## v0.12.0
 
 - Added support for Populating Cloud RoleName, RoleInstance from Resource's
