+++ opentelemetry-user-events-logs/benches/logs.rs
@@ -29,14 +29,13 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use opentelemetry_appender_tracing::layer as tracing_layer;
 use opentelemetry_sdk::logs::SdkLoggerProvider;
 use opentelemetry_sdk::Resource;
-use opentelemetry_user_events_logs::{build_processor, ExportOptions};
+use opentelemetry_user_events_logs::Processor;
 use tracing::error;
 use tracing_subscriber::prelude::*;
 use tracing_subscriber::Registry;
 
 fn benchmark_with_ot_layer(c: &mut Criterion, name: &str, num_attributes: usize) {
-    let export_options = ExportOptions::builder("myprovider").build().unwrap();
-    let user_event_processor = build_processor(export_options);
+    let user_event_processor = Processor::builder("myprovider").build().unwrap();
     let provider = SdkLoggerProvider::builder()
         .with_resource(
             Resource::builder_empty()
