+++ opentelemetry-user-events-logs/src/logs/exporter.rs
@@ -11,84 +11,6 @@ use std::{cell::RefCell, str, time::SystemTime};
 
 thread_local! { static EBW: RefCell<EventBuilder> = RefCell::new(EventBuilder::new());}
 
-/// Options for configuring the User Events Exporter.
-#[derive(Debug)]
-pub struct ExportOptions<'a> {
-    provider_name: &'a str,
-}
-
-impl<'a> ExportOptions<'a> {
-    /// Returns the provider name.
-    fn provider_name(&self) -> &'a str {
-        self.provider_name
-    }
-
-    /// Builder for `EventsOptions`, with a mandatory provider_name.
-    /// This is the name of the provider that will be used to register the event sets.
-    /// The provider name must:
-    /// - Not be empty.
-    /// - Be less than 234 characters.
-    /// - Contain only ASCII letters, digits, and the underscore (`'_'`) character.
-    /// - Be short, human-readable, and unique enough to avoid conflicts with other provider names.
-    /// - Typically include a company name and a component name, e.g., "MyCompany_MyComponent".
-    ///
-    /// If an invalid provider name is provided, this method will not add the exporter to the builder.
-    ///
-    /// Tracepoint names are generated by combining the provider name, event
-    /// level and keyword (currently hardcoded to `1`) in the following format:
-    /// `ProviderName + '_' + 'L' + EventLevel + 'K' + EventKeyword`
-    ///
-    /// For example, if "myprovider" is the provider name, the following tracepoint names are created:
-    /// - `myprovider_L5K1`
-    /// - `myprovider_L4K1`
-    /// - `myprovider_L3K1`
-    /// - `myprovider_L2K1`
-    /// - `myprovider_L1K1`
-    ///
-    /// perf tool can be used to record events from the tracepoints.
-    /// For example the following will capture level 2 (Error) and 3(Warning) events:
-    /// perf record -e user_events:myprovider_L2K1,user_events:myprovider_L3K1
-    pub fn builder(provider_name: &'a str) -> ExportOptionsBuilder<'a> {
-        ExportOptionsBuilder::new(provider_name)
-    }
-}
-
-/// Builder for `ExportOptions`.
-#[derive(Debug)]
-pub struct ExportOptionsBuilder<'a> {
-    provider_name: &'a str,
-}
-
-impl<'a> ExportOptionsBuilder<'a> {
-    /// Creates a new builder with a mandatory provider_name.
-    fn new(provider_name: &'a str) -> Self {
-        Self { provider_name }
-    }
-
-    /// Builds the `ExportOptions` or returns an error if validation fails.
-    pub fn build(self) -> Result<ExportOptions<'a>, String> {
-        let provider_name = self.provider_name;
-
-        // Validate provider_name
-        if provider_name.is_empty() {
-            return Err("Provider name cannot be empty.".to_string());
-        }
-        if provider_name.len() >= 234 {
-            return Err("Provider name must be less than 234 characters.".to_string());
-        }
-        if !provider_name
-            .chars()
-            .all(|c| c.is_ascii_alphanumeric() || c == '_')
-        {
-            return Err(
-                "Provider name must contain only ASCII letters, digits, and '_'.".to_string(),
-            );
-        }
-
-        Ok(ExportOptions { provider_name })
-    }
-}
-
 /// UserEventsExporter is a log exporter that exports logs in EventHeader format to user_events tracepoint.
 pub(crate) struct UserEventsExporter {
     provider: Mutex<Provider>,
@@ -102,8 +24,7 @@ const EVENT_ID: &str = "event_id";
 
 impl UserEventsExporter {
     /// Create instance of the exporter
-    pub(crate) fn new(options: ExportOptions) -> Self {
-        let provider_name = options.provider_name();
+    pub(crate) fn new(provider_name: &str) -> Self {
         let mut eventheader_provider: Provider =
             Provider::new(provider_name, &Provider::new_options());
         let event_sets = Self::register_events(&mut eventheader_provider);
@@ -221,6 +142,12 @@ impl UserEventsExporter {
         }
     }
 
+    /// Gets the event name from the log record
+    fn get_event_name<'a>(&self, _record: &'a opentelemetry_sdk::logs::SdkLogRecord) -> &'a str {
+        // TODO: Add callback to get event name from the log record
+        "Log"
+    }
+
     pub(crate) fn export_log_data(
         &self,
         log_record: &opentelemetry_sdk::logs::SdkLogRecord,
@@ -256,11 +183,8 @@ impl UserEventsExporter {
                 // for each event.
                 // TODO: What if the event name is not provided? "Log" is used as default.
                 // TODO: Should event_tag be non-zero?
-                let event_name = log_record
-                    .event_name()
-                    .filter(|s| !s.trim().is_empty())
-                    .unwrap_or("Log");                
-                eb.reset("Log", 0);
+                let event_name = self.get_event_name(log_record);
+                eb.reset(event_name, 0);
 
                 eb.add_value("__csver__", 1024, FieldFormat::UnsignedInt, 0); // 0x400 in hex
 
@@ -381,10 +305,14 @@ impl UserEventsExporter {
                     eb.add_value("eventId", event_id, FieldFormat::SignedInt, 0);
                     cs_b_count += 1;
                 }
-                // TODO: eventname is already added to header.
-                // Should we add it again?
-                // Or should we use Target?
-                eb.add_str("name", event_name, FieldFormat::Default, 0);
+
+                let partb_name = log_record
+                    .event_name()
+                    .filter(|s| !s.trim().is_empty());
+                if let Some(name) = partb_name {
+                    eb.add_str("name", name, FieldFormat::Default, 0);
+                }
+
                 cs_b_count += 1;
                 eb.set_struct_field_count(cs_b_bookmark, cs_b_count);
 
@@ -472,89 +400,10 @@ mod tests {
     use super::*;
     #[test]
     fn exporter_debug() {
-        let options = ExportOptions::builder("test_provider")
-            .build()
-            .expect("Failed to create ExportOptions");
-        let exporter = UserEventsExporter::new(options);
+        let exporter = UserEventsExporter::new("test_provider");
         assert_eq!(
             format!("{:?}", exporter),
             "user_events log exporter (provider name: test_provider)"
         );
-    }
-
-    #[test]
-    fn valid_provider_name() {
-        let valid_names = vec![
-            "ValidName",
-            "valid_name",
-            "Valid123",
-            "valid_123",
-            "_valid_name",
-            "VALID_NAME",
-        ];
-
-        for valid_name in valid_names {
-            let options = ExportOptions::builder(valid_name)
-                .build()
-                .expect("Failed to build ExportOptions");
-            let _exporter = UserEventsExporter::new(options);
-        }
-    }
-
-    #[test]
-    fn provider_name_too_long() {
-        let long_name = "a".repeat(234);
-        let options = ExportOptions::builder(&long_name).build();
-        assert!(options.is_err());
-        assert_eq!(
-            options.err().unwrap(),
-            "Provider name must be less than 234 characters.".to_string()
-        );
-    }
-
-    #[test]
-    fn provider_name_contains_invalid_characters() {
-        // Define a vector of invalid provider names to test
-        let invalid_names = vec![
-            "Invalid Name",  // space
-            "Invalid:Name",  // colon
-            "Invalid\0Name", // null character
-            "Invalid-Name",  // hyphen
-            "InvalidName!",  // exclamation mark
-            "InvalidName@",  // at symbol
-            "Invalid+Name",  // plus
-            "Invalid&Name",  // ampersand
-            "Invalid#Name",  // hash
-            "Invalid%Name",  // percent
-            "Invalid/Name",  // slash
-            "Invalid\\Name", // backslash
-            "Invalid=Name",  // equals
-            "Invalid?Name",  // question mark
-            "Invalid;Name",  // semicolon
-            "Invalid,Name",  // comma
-        ];
-
-        // Expected error message
-        let expected_error =
-            "Provider name must contain only ASCII letters, digits, and '_'.".to_string();
-
-        // Test each invalid name
-        for invalid_name in invalid_names {
-            let options = ExportOptions::builder(invalid_name).build();
-            // Assert that the result is an error
-            assert!(
-                options.is_err(),
-                "Expected '{}' to be invalid, but it was accepted",
-                invalid_name
-            );
-
-            // Assert that the error message is as expected
-            assert_eq!(
-                options.err().unwrap(),
-                expected_error,
-                "Wrong error message for invalid name: '{}'",
-                invalid_name
-            );
-        }
     }
 }
